requirements.txt
torch>=2.1
torchvision>=0.16
torchaudio>=2.1
timm>=0.9.16
segmentation-models-pytorch>=0.3.3
transformers>=4.44
datasets>=2.20
opencv-python>=4.9
albumentations>=1.4.7
pyyaml>=6.0.2
numpy>=1.26
scipy>=1.11
tqdm>=4.66
pillow>=10.3
scikit-image>=0.24
